Vise info om objekt (#26)
The app now recognizes a barcode symbol (number 0), and displays a hardcoded text when the barcode is recognized.

The barcode is generated through https://au.gmented.com/app/marker/marker.php with 3x3 matrix dimensions. The code used for generation is simply: 0. No data is encoded in the symbol, we can only recognize it. The text is hardcoded in the html and can be changed to fit whatever object we want to recognize later.