Kalman filter walking (#112)
Added a kalman-gps-camera component. Adding a kalman filter to the pre-existing component gps-camera.

The point is to provide a more stable camera component that works when walking around in a location-based AR experience. How it works is described in developer-guide.md

Tested by jonas and ola. Seems a bit more stable, but more testing and tuning is needed.